replace build_imports
